fix(clippy): Resolve all 65 warnings across openlife-core
Auto-fixed via cargo clippy --fix:
- Redundant field names, closures, to_string in format!
- map().flatten() -> and_then()
- or_insert_with -> or_default()
- Derivable Default implementations
- Collapsible if statements
- Simplified boolean expressions

Manual fixes:
- Removed unused ExtractedGoal.confidence field
- Removed unused any_critical_failure variable
- Added clamp() for min/max patterns
- Added strip_prefix() for manual stripping
- Added sort_by_key() for sort_by
- Created TypedPlan type alias for complex type
- Added #[allow(too_many_arguments)] for internal helpers
- Used struct init with ..Default::default()

Verification:
- cargo clippy -- -D warnings: 0 errors
- cargo test: 187 passed, 0 failed

Refs: Task 2, clippy cleanup

+++ openlife-core/src/feedback.rs
@@ -633,7 +633,7 @@ impl FeedbackStore {
             .conn
             .lock()
             .map_err(|e| anyhow::anyhow!("mutex poison: {}", e))?;
-        let cutoff = (chrono::Utc::now() - chrono::Duration::days(period_days as i64)).to_rfc3339();
+        let cutoff = (chrono::Utc::now() - chrono::Duration::days(period_days)).to_rfc3339();
 
         let feedback_up: i64 = conn
             .query_row(
